+++ lib/generators/shopify_app/install/templates/user_agent.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-module ShopifyAPI
-  class Base < ActiveResource::Base
-    headers['User-Agent'] << " | ShopifyApp/#{ShopifyApp::VERSION}"
-  end
-end