+++ test/app_templates/config/application.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
-require File.expand_path('../boot', __FILE__)
 
-require 'rails/all'
+require File.expand_path("../boot", __FILE__)
+
+require "rails/all"
 
 Bundler.require(*Rails.groups)