+++ test/generators/routes_generator_test.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
-require 'test_helper'
-require 'generators/shopify_app/routes/routes_generator'
+
+require "test_helper"
+require "generators/shopify_app/routes/routes_generator"
 
 class ControllerGeneratorTest < Rails::Generators::TestCase
   tests ShopifyApp::Generators::RoutesGenerator
@@ -15,10 +16,10 @@ class ControllerGeneratorTest < Rails::Generators::TestCase
     run_generator
 
     assert_file "config/routes.rb" do |routes|
-      assert_match "get 'login' => :new, :as => :login", routes
-      assert_match "post 'login' => :create, :as => :authenticate", routes
-      assert_match "get 'auth/shopify/callback' => :callback", routes
-      assert_match "get 'logout' => :destroy, :as => :logout", routes
+      assert_match "get \"login\" => :new, :as => :login", routes
+      assert_match "post \"login\" => :create, :as => :authenticate", routes
+      assert_match "get \"auth/shopify/callback\" => :callback", routes
+      assert_match "get \"logout\" => :destroy, :as => :logout", routes
       refute_match "mount ShopifyApp::Engine, at: '/'", routes
     end
   end